Refactor vector class to make it more general. Now I will create geometric vector & color.
Refactor vector class with using the rule of three(I am not declaring copy constructor, assign operator and destructor, so it will be created automatically and I approve it.
Changed getter functions names from x() to getX() to increase clarity.
Vector will be a general class kernel, and geometric vector and color will be inherited from it.